Updates to get both workers functional
