+++ worker_go/scripts/start-worker.sh
