+++ worker_go/worker/main.go
@@ -5,23 +5,30 @@ package main
 
 import (
 	"log"
+	"os"
 
 	"go.temporal.io/sdk/client"
-	"go.temporal.io/sdk/contrib/envconfig"
 	"go.temporal.io/sdk/worker"
 
 	"github.com/canonical/charmed-temporal-uats/helloworld"
 )
 
 func main() {
+	host := os.Getenv("TEMPORAL_HOST")
+	namespace := os.Getenv("TEMPORAL_NAMESPACE")
+	taskQueue := os.Getenv("TEMPORAL_QUEUE")
+
 	// The client and worker are heavyweight objects that should be created once per process.
-	c, err := client.Dial(envconfig.MustLoadDefaultClientOptions())
+	c, err := client.Dial(client.Options{
+		HostPort: host,
+		Namespace: namespace,
+	})
 	if err != nil {
 		log.Fatalln("Unable to create client", err)
 	}
 	defer c.Close()
 
-	w := worker.New(c, "hello-world", worker.Options{})
+	w := worker.New(c, taskQueue, worker.Options{})
 
 	w.RegisterWorkflow(helloworld.Workflow)
 	w.RegisterActivity(helloworld.Activity)
