+++ worker_python/scripts/start-worker.sh
