--- worker_python/src/activities/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Sample Temporal activities."""
