--- worker_python/src/workflows/__init__.py
@@ -1,4 +1,0 @@
-# Copyright 2025 Canonical Ltd.
-# See LICENSE file for licensing details.
-
-"""Sample Temporal workflows."""
